requirements_long_haul.txt
anyio==4.3.0
blinker==1.7.0
click==8.1.7
dnspython==1.16.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
graphql-core==3.2.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pymongo==3.11.0
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.37.1
typing_extensions==4.10.0
Werkzeug==3.0.1
modal==1.14.1
langchain_openai==0.0.8
modal==0.61.88
torch==2.2.1+cu118
transformers==4.38.2